hw3.v2

+++ msu_spring_2019/C++/03/matrix.h
@@ -51,11 +51,11 @@ public:
            throw std::out_of_range("");
     }
 
-    int getRows()
+    size_t getRows()
     {
         return (rows);
     }
-    int getColumns()
+    size_t getColumns()
     {
         return (cols);
     }
